Project Yatra/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0 
colorama==0.4.6 
comm==0.1.2     
debugpy==1.6.4  
decorator==5.1.1
distlib==0.3.6  
Django==4.1.5   
django-cors-headers==3.13.0
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==22.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==24.0.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
virtualenv==20.19.0
wcwidth==0.2.5